Added house jobs list and cleaning some code
